+++ atoml/cross_validation/hierarchy_cv.py
@@ -1,7 +1,6 @@
 """Cross validation routines to work with feature database."""
 import sqlite3
 import json
-import yaml
 import pickle
 import numpy as np
 from random import shuffle
@@ -110,17 +109,17 @@ class Hierarchy(object):
 
     def load_split(self):
         """Function to load the split from file."""
-        if self.file_format is not 'pickle':
-            with open(self.file_name + '.' +
-                      self.file_format, 'r') as textfile:
-                if self.file_format is 'json':
+        if self.file_format is 'json':
+            with open('{0}.{1}'.format(self.file_name, self.file_format),
+                      'r') as textfile:
                     data = json.load(textfile)
-                if self.file_format is 'yaml':
-                    data = yaml.load(textfile)
-        else:
-            with open(self.file_name + '.' +
-                      self.file_format, 'rb') as textfile:
+        elif self.file_format is 'pickle':
+            with open('{0}.{1}'.format(self.file_name, self.file_format),
+                      'rb') as textfile:
                 data = pickle.load(textfile)
+        else:
+            raise NotImplementedError(
+                '{} format not supported'.format(self.file_format))
 
         return data
 
@@ -253,17 +252,17 @@ class Hierarchy(object):
         data : dict
             Index dict generated within the split_index function.
         """
-        if self.file_format is not 'pickle':
-            with open(self.file_name + '.' +
-                      self.file_format, 'w') as textfile:
-                if self.file_format is 'json':
+        if self.file_format == 'json':
+            with open('{0}.{1}'.format(self.file_name, self.file_format),
+                      'w') as textfile:
                     json.dump(data, textfile)
-                if self.file_format is 'yaml':
-                    yaml.dump(data, textfile)
-        else:
-            with open(self.file_name + '.' +
-                      self.file_format, 'wb') as textfile:
+        elif self.file_format == 'pickle':
+            with open('{0}.{1}'.format(self.file_name, self.file_format),
+                      'wb') as textfile:
                 pickle.dump(data, textfile, protocol=pickle.HIGHEST_PROTOCOL)
+        else:
+            raise NotImplementedError(
+                '{} format not supported'.format(self.file_format))
 
     def _compile_split(self, id_list):
         """Function to get actual data from database.